requirements-api.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
PyYAML>=6.0
pydantic>=2.0.0
numpy>=1.24.0
redis>=5.0.0
async-timeout>=4.0.0
mss>=9.0.0
Pillow>=10.0.0
